Reorganizando todos los componetes
