+++ frontend-law/src/components/ui/Dashboard.jsx
@@ -1,24 +1,15 @@
-/*
-  This example requires some changes to your config:
-  
-  ```
-  // tailwind.config.js
-  module.exports = {
-    // ...
-    plugins: [
-      // ...
-      require('@tailwindcss/forms'),
-    ],
-  }
-  ```
-*/
 import { Fragment, useState } from 'react'
 import { Dialog, Disclosure, Menu, Transition } from '@headlessui/react'
 import { XMarkIcon } from '@heroicons/react/24/outline'
 import { ChevronDownIcon, MinusIcon, PlusIcon, Squares2X2Icon, Bars3Icon } from '@heroicons/react/20/solid'
-import { Area, AreaChart, Bar, BarChart, CartesianGrid, Cell, Legend, Pie, PieChart, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts'
 import logo from '../../assets/logo.png'
 import DataTableResults from './DataTableResults'
+import Form from './Form'
+import GrafStatGroup from './GrafStatGroup'
+import InfoCardsGroup from './InfoCardsGroup'
+import GreetingForPerson from './GreetingForPerson'
+import LeftMenuDash from './LeftMenuDash'
+import BackgroundImage from './BackgroundImage'
 
 const sortOptions = [
     { name: 'Most Popular', href: '#', current: true },
@@ -71,30 +62,6 @@ const subMenus = [
         ],
     },
 ]
-const stats = [
-    { id: 1, titulo: 'Weekly', valor: '714k', icono: 'https://minimal-kit-react.vercel.app/assets/icons/glass/ic_glass_bag.png' },
-    { id: 2, titulo: 'New Users', valor: '1.35m', icono: 'https://minimal-kit-react.vercel.app/assets/icons/glass/ic_glass_users.png' },
-    { id: 3, titulo: 'Item Orders', valor: '1.72m', icono: 'https://minimal-kit-react.vercel.app/assets/icons/glass/ic_glass_buy.png' },
-    { id: 4, titulo: 'Bug Reports', valor: '234', icono: 'https://minimal-kit-react.vercel.app/assets/icons/glass/ic_glass_message.png' }
-]
-const data = [
-    {name: "María", age: 10, weight: 60},
-    {name: 'Karina', age: 25, weight: 70},
-    {name: 'Susana', age: 15, weight: 65},
-    {name: 'Pedro', age: 35, weight: 85},
-    {name: 'Felipe', age: 12, weight: 48},
-    {name: 'Laura', age: 30, weight: 69},
-    {name: 'Adrián', age: 15, weight: 78},
-]
-const dataPay = [
-    { name: "Group A", value: 2400 },
-    { name: 'Group B', value: 4567 },
-    { name: 'Group C', value: 1398 },
-    { name: 'Group D', value: 9800 },
-    { name: 'Group E', value: 3908 },
-    { name: 'Group F', value: 4800 },
-]
-const COLORS = ['#ce93d8', '#5c6bc0', '#b39ddb', '#4dd0e1', '#f48fb1', '#d500f9']
 
 function classNames(...classes) {
     return classes.filter(Boolean).join(' ')
@@ -105,36 +72,7 @@ export default function Dashboard() {
 
     return (
         <div className="relative isolate overflow-hidden bg-gray-900">
-            <img
-                src="https://images.unsplash.com/photo-1521737604893-d14cc237f11d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&crop=focalpoint&fp-y=.8&w=2830&h=1500&q=80&blend=111827&sat=-100&exp=15&blend-mode=multiply"
-                alt=""
-                className="absolute inset-0 -z-10 h-full w-full object-cover object-right md:object-center blur-sm"
-            />
-            <div
-                className="hidden sm:absolute sm:-top-10 sm:right-1/2 sm:-z-10 sm:mr-10 sm:block sm:transform-gpu sm:blur-3xl"
-                aria-hidden="true"
-            >
-                <div
-                    className="aspect-[1097/845] w-[68.5625rem] bg-gradient-to-tr from-[#ff4694] to-[#776fff] opacity-20"
-                    style={{
-                        clipPath:
-                            'polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)',
-                    }}
-                />
-            </div>
-            
-            <div
-                className="absolute -top-52 left-1/2 -z-10 -translate-x-1/2 transform-gpu blur-3xl sm:top-[-28rem] sm:ml-16 sm:translate-x-0 sm:transform-gpu"
-                aria-hidden="true">
-
-                <div
-                    className="aspect-[1097/845] w-[68.5625rem] bg-gradient-to-tr from-[#ff4694] to-[#776fff] opacity-20"
-                    style={{
-                        clipPath:
-                            'polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)',
-                    }}
-                />
-            </div>
+            <BackgroundImage />
 
             <div>
                 {/* Mobile filter dialog */}
@@ -147,9 +85,9 @@ export default function Dashboard() {
                             enterTo="opacity-100"
                             leave="transition-opacity ease-linear duration-300"
                             leaveFrom="opacity-100"
-                            leaveTo="opacity-0"
-                        >
-                        <div className="fixed inset-0 bg-black bg-opacity-25" />
+                            leaveTo="opacity-0" >
+
+                            <div className="fixed inset-0 bg-black bg-opacity-25" />
                         </Transition.Child>
 
                         <div className="fixed inset-0 z-40 flex">
@@ -160,8 +98,8 @@ export default function Dashboard() {
                                 enterTo="translate-x-0"
                                 leave="transition ease-in-out duration-300 transform"
                                 leaveFrom="translate-x-0"
-                                leaveTo="translate-x-full"
-                            >
+                                leaveTo="translate-x-full" >
+
                                 <Dialog.Panel className="relative ml-auto flex h-full w-full max-w-xs flex-col overflow-y-auto bg-white py-4 pb-12 shadow-xl">
                                     <div className="flex items-center justify-between px-4">
                                         <h2 className="text-lg font-medium text-gray-900">Diego</h2>
@@ -234,6 +172,7 @@ export default function Dashboard() {
                                 </Dialog.Panel>
                             </Transition.Child>
                         </div>
+
                     </Dialog>
                 </Transition.Root>
 
@@ -248,7 +187,6 @@ export default function Dashboard() {
                         </div>
 
                         <div className="flex items-center">
-
                             <Menu as="div" className="relative inline-block text-left">
                                 <div>
                                     <Menu.Button className="group inline-flex justify-center text-sm font-medium text-white hover:text-[#E37B3A]">
@@ -267,8 +205,8 @@ export default function Dashboard() {
                                     enterTo="transform opacity-100 scale-100"
                                     leave="transition ease-in duration-75"
                                     leaveFrom="transform opacity-100 scale-100"
-                                    leaveTo="transform opacity-0 scale-95"
-                                >
+                                    leaveTo="transform opacity-0 scale-95" >
+
                                     <Menu.Items className="absolute right-0 z-10 mt-2 w-40 origin-top-right rounded-md bg-white shadow-2xl ring-1 ring-black ring-opacity-5 focus:outline-none">
                                         <div className="py-1">
                                             {sortOptions.map((option) => (
@@ -290,16 +228,16 @@ export default function Dashboard() {
                                         </div>
                                     </Menu.Items>
                                 </Transition>
+
                             </Menu>
 
                             <button type="button" className="-m-2 ml-5 p-2 text-white hover:text-[#E37B3A] sm:ml-7">
                                 <span className="sr-only">View grid</span>
                                 <Squares2X2Icon className="h-5 w-5" aria-hidden="true" />
                             </button>
-                            
+
                             <button type="button" className="-m-2 ml-4 p-2 text-white hover:text-[#E37B3A] sm:ml-6 lg:hidden"
                                 onClick={() => setMobileFiltersOpen(true)}>
-
                                 <span className="sr-only">Filters</span>
                                 <Bars3Icon className="h-5 w-5" aria-hidden="true" />
                             </button>
@@ -311,315 +249,19 @@ export default function Dashboard() {
                     <section aria-labelledby="products-heading" className="pb-6 pt-6">
                         <div className="grid grid-cols-1 p-2 gap-x-8 gap-y-10 lg:grid-cols-4">
                             {/* Filters */}
-                            <form className="hidden lg:block">
-                                <div className='bg-[#100C18] text-base text-white h-full border border-white rounded-2xl px-3 py-3'>
-                                    <ul role="list" className="space-y-4 border-b border-gray-200 pb-6 font-medium">
-                                        {subCategories.map((category) => (
-                                            <li key={category.name}>
-                                                <a href={category.href}>{category.name}</a>
-                                            </li>
-                                        ))}
-                                    </ul>
-
-                                    {subMenus.map((section) => (
-
-                                        <Disclosure as="div" key={section.id} className="border-b border-gray-200 py-6">
-                                            {({ open }) => (
-                                                <>
-                                                    <h3 className="-my-3 flow-root">
-                                                        <Disclosure.Button className="flex w-full items-center justify-between bg-white py-3 text-sm text-gray-400 hover:text-gray-500">
-                                                            <span className="font-medium text-gray-900">{section.name}</span>
-                                                            <span className="ml-6 flex items-center">
-                                                                {open ? (
-                                                                    <MinusIcon className="h-5 w-5" aria-hidden="true" />
-                                                                ) : (
-                                                                    <PlusIcon className="h-5 w-5" aria-hidden="true" />
-                                                                )}
-                                                            </span>
-                                                        </Disclosure.Button>
-                                                    </h3>
-                                                    <Disclosure.Panel className="pt-6">
-                                                        <div className="space-y-4">
-                                                            {section.options.map((option, optionIdx) => (
-                                                                <div key={option.value} className="flex items-center">
-                                                                    <input
-                                                                        id={`filter-${section.id}-${optionIdx}`}
-                                                                        name={`${section.id}[]`}
-                                                                        defaultValue={option.value}
-                                                                        type="checkbox"
-                                                                        defaultChecked={option.checked}
-                                                                        className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500"
-                                                                    />
-                                                                    <label
-                                                                        htmlFor={`filter-${section.id}-${optionIdx}`}
-                                                                        className="ml-3 text-sm text-white"
-                                                                    >
-                                                                        {option.label}
-                                                                    </label>
-                                                                </div>
-                                                            ))}
-                                                        </div>
-                                                    </Disclosure.Panel>
-                                                </>
-                                            )}
-                                        </Disclosure>
-
-                                    ))}
-                                </div>
-                            </form>
+                            <LeftMenuDash />
 
                             {/* Product grid */}
                             {/*INSPIRACION: https://mui.com/store/previews/minimal-dashboard-free/ */}
                             <div className="lg:col-span-3 py-2">{/* Your content */}
                                 {/*Mensaje de bienvenida*/}
-                                <div className="mx-auto max-w-7xl px-10 lg:px-8">
-                                    <h3 className='text-2xl text-white'>Bienvenido Dr. Chapatin</h3>
-                                </div>
-
+                                <GreetingForPerson />
                                 {/*Contenedor para mis etiquetas */}
-                                <div className="flex flex-row justify-around flex-wrap py-5 px-6 text-white">
-                                    {/*Para mis tarjetas etiquetas */}
-                                    {stats.map((stat) => (
-                                        <div key={stat.id} className="flex flex-row justify-start backdrop-blur-3xl border border-white rounded-2xl w-48 h-24 my-2 mx-2">
-                                            <div className='px-2 py-2'>
-                                                <img src={stat.icono} alt="icono-card" />
-                                            </div>
-                                            <div className='flex flex-col justify-center'>
-                                                <h1 className='text-2xl'>{stat.valor}</h1>
-                                                <h1>{stat.titulo}</h1>
-                                            </div>
-                                        </div>
-                                    ))}
-
-                                </div>
-
+                                <InfoCardsGroup />
                                 {/*Contenedor para mis graficos estadisticos */}
-                                <div className="mx-auto max-w-7xl px-6 lg:px-8">
-                                    <h3 className='text-2xl text-white pt-4 pl-2'>Bienvenido Dr. Chapatin</h3>
-                                    <div className="sm:mx-auto flex flex-row justify-start border border-white rounded-2xl h-72 my-2 mx-2 pt-16 sm:pt-0">
-                                        <ResponsiveContainer width="100%" aspect={2}>
-                                            <BarChart
-                                                data={data}
-                                                width={500}
-                                                height={400}
-                                                margin={{
-                                                    top: 10,
-                                                    right: 30,
-                                                    left: 0,
-                                                    bottom: 0
-                                                }}
-                                            >
-                                                <CartesianGrid strokeDasharray="4 1 2" />
-                                                <XAxis dataKey="name" />
-                                                <YAxis />
-                                                <Tooltip />
-                                                <Legend />
-                                                <Bar dataKey="weight" fill="#6b48ff" />
-                                                <Bar dataKey="age" fill="#1ee3cf" />
-                                            </BarChart>
-                                        </ResponsiveContainer>
-                                    </div>
-
-                                    <h3 className='text-2xl text-white pt-4 pl-2'>Bienvenido Dr. Chapatin</h3>
-                                    <div className="sm:mx-auto flex flex-row justify-start border border-white rounded-2xl h-72 my-2 mx-2 pt-16 sm:pt-0">
-                                        <ResponsiveContainer width="100%" aspect={2}>
-                                            <AreaChart
-                                                width={500}
-                                                height={400}
-                                                data={data}
-                                                margin={{
-                                                    top: 10,
-                                                    right: 30,
-                                                    left: 0,
-                                                    bottom: 0
-                                                }}
-                                            >
-                                                <CartesianGrid strokeDasharray="3 3" />
-                                                <XAxis dataKey="name" />
-                                                <YAxis />
-                                                <Tooltip />
-                                                <Area type="monotone" dataKey="age" stackId="1" stroke='#8884d8' fill="#8884d8" />
-                                                <Area type="monotone" dataKey="weight" stackId="1" stroke='#82caed' fill="#fad3cf" />
-                                            </AreaChart>
-                                        </ResponsiveContainer>
-                                    </div>
-
-                                    <h3 className='text-2xl text-white pt-4 pl-2'>Bienvenido Dr. Chapatin</h3>
-                                    <div className="sm:mx-auto flex flex-row justify-start border border-white rounded-2xl h-72 my-2 mx-2">
-                                        <ResponsiveContainer>
-                                            <PieChart>
-                                                <Pie
-                                                    dataKey="value"
-                                                    data={dataPay}
-                                                    innerRadius={60}
-                                                    outerRadius={120}
-                                                    fill="#82ca9d"
-                                                >
-                                                    {dataPay.map((entry, index) => (
-                                                        <Cell key={`cell-${index}`} fill={COLORS[index % COLORS.length]} />
-                                                    ))}
-                                                </Pie>
-                                                <Tooltip />
-                                            </PieChart>
-                                        </ResponsiveContainer>
-                                    </div>
-                                </div>
-   
-
-                                <div className=''>
-                                    <h3 className='text-2xl text-white pt-4 pl-2 pb-2'>Formulario</h3>
-                                    <div className='border border-white rounded-2xl p-8'>
-                                        <form method="POST" action="#">
-                                            <div className="space-y-12">
-                                                <div className="border-b border-gray-900/10 pb-12">
-                                                    <h2 className="text-base font-semibold leading-7 text-white">Personal Information</h2>
-                                                    <p className="mt-1 text-sm leading-6 text-white">Use a permanent address where you can receive mail.</p>
-
-                                                    <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
-                                                        <div className="sm:col-span-3">
-                                                            <label htmlFor="first-name" className="block text-sm font-medium leading-6 text-white">
-                                                                First name
-                                                            </label>
-                                                            <div className="mt-2">
-                                                                <input
-                                                                    type="text"
-                                                                    name="first-name"
-                                                                    id="first-name"
-                                                                    autoComplete="given-name"
-                                                                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                                                />
-                                                            </div>
-                                                        </div>
-
-                                                        <div className="sm:col-span-3">
-                                                            <label htmlFor="last-name" className="block text-sm font-medium leading-6 text-white">
-                                                                Last name
-                                                            </label>
-                                                            <div className="mt-2">
-                                                                <input
-                                                                    type="text"
-                                                                    name="last-name"
-                                                                    id="last-name"
-                                                                    autoComplete="family-name"
-                                                                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                                                />
-                                                            </div>
-                                                        </div>
-
-                                                        <div className="sm:col-span-4">
-                                                            <label htmlFor="email" className="block text-sm font-medium leading-6 text-white">
-                                                                Email address
-                                                            </label>
-                                                            <div className="mt-2">
-                                                                <input
-                                                                    id="email"
-                                                                    name="email"
-                                                                    type="email"
-                                                                    autoComplete="email"
-                                                                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                                                />
-                                                            </div>
-                                                        </div>
-
-                                                        <div className="sm:col-span-3">
-                                                            <label htmlFor="country" className="block text-sm font-medium leading-6 text-white">
-                                                                Country
-                                                            </label>
-                                                            <div className="mt-2">
-                                                                <select
-                                                                    id="country"
-                                                                    name="country"
-                                                                    autoComplete="country-name"
-                                                                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:max-w-xs sm:text-sm sm:leading-6"
-                                                                >
-                                                                    <option>United States</option>
-                                                                    <option>Canada</option>
-                                                                    <option>Mexico</option>
-                                                                </select>
-                                                            </div>
-                                                        </div>
-
-                                                        <div className="col-span-full">
-                                                            <label htmlFor="street-address" className="block text-sm font-medium leading-6 text-white">
-                                                                Street address
-                                                            </label>
-                                                            <div className="mt-2">
-                                                                <input
-                                                                    type="text"
-                                                                    name="street-address"
-                                                                    id="street-address"
-                                                                    autoComplete="street-address"
-                                                                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                                                />
-                                                            </div>
-                                                        </div>
-
-                                                        <div className="sm:col-span-2 sm:col-start-1">
-                                                            <label htmlFor="city" className="block text-sm font-medium leading-6 text-white">
-                                                                City
-                                                            </label>
-                                                            <div className="mt-2">
-                                                                <input
-                                                                    type="text"
-                                                                    name="city"
-                                                                    id="city"
-                                                                    autoComplete="address-level2"
-                                                                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                                                />
-                                                            </div>
-                                                        </div>
-
-                                                        <div className="sm:col-span-2">
-                                                            <label htmlFor="region" className="block text-sm font-medium leading-6 text-white">
-                                                                State / Province
-                                                            </label>
-                                                            <div className="mt-2">
-                                                                <input
-                                                                    type="text"
-                                                                    name="region"
-                                                                    id="region"
-                                                                    autoComplete="address-level1"
-                                                                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                                                />
-                                                            </div>
-                                                        </div>
-
-                                                        <div className="sm:col-span-2">
-                                                            <label htmlFor="postal-code" className="block text-sm font-medium leading-6 text-white">
-                                                                ZIP / Postal code
-                                                            </label>
-                                                            <div className="mt-2">
-                                                                <input
-                                                                    type="text"
-                                                                    name="postal-code"
-                                                                    id="postal-code"
-                                                                    autoComplete="postal-code"
-                                                                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-                                                                />
-                                                            </div>
-                                                        </div>
-                                                    </div>
-                                                </div>
-                                            </div>
-
-                                            <div className="mt-6 flex items-center justify-end gap-x-6">
-                                                <button type="button" className="text-sm font-semibold leading-6 text-white">
-                                                    Cancel
-                                                </button>
-                                                <button
-                                                    type="submit"
-                                                    className="rounded-md bg-orange-500 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-blue-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-                                                >
-                                                    Save
-                                                </button>
-                                            </div>
-                                        </form>
-                                    </div>
-                                </div>
-
-
+                                <GrafStatGroup />
+                                <Form />
                                 <DataTableResults />
-
                             </div>
 
                         </div>
