+++ frontend-law/src/components/ui/DashboardPath.jsx
@@ -1,0 +1,9 @@
+import Dashboard from "./Dashboard";
+
+export default function DashboardPath() {
+    return(
+        <div>
+            <Dashboard />
+        </div>
+    )
+}
