+++ frontend-law/src/components/ui/GreetingForPerson.jsx
@@ -1,0 +1,7 @@
+export default function GreetingForPerson() {
+    return (
+        <div className="mx-auto max-w-7xl px-10 lg:px-8">
+            <h3 className='text-2xl text-white'>Bienvenido Dr. Chapatin</h3>
+        </div>
+    )
+}
