+++ frontend-law/src/components/ui/LoginPath.jsx
@@ -1,0 +1,13 @@
+import { FooterWithSocialLinks } from "./FooterWithSocialLinks";
+import Header from "./Header";
+import Login from "./Login";
+
+export default function LoginPath() {
+    return(
+        <div>
+            <Header />
+            <Login />
+            <FooterWithSocialLinks />
+        </div>
+    )
+}
